+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## v0.20
 
-
-
+* added internet type ip
+* added internet type ipv4
+* added internet type ipv6
+* added internet type email
 
 ## v0.10